Crate bitgrin_wallet_api

Higher level wallet functions which can be used by callers to operate on the wallet, as well as helpers to invoke and instantiate wallets and listeners

Structs

Foreign
Main interface into all wallet API functions. Wallet APIs are split into two seperate blocks of functionality called the 'Owner' and 'Foreign' APIs

Owner
Main interface into all wallet API functions. Wallet APIs are split into two seperate blocks of functionality called the 'Owner' and 'Foreign' APIs

Enums

ForeignCheckMiddlewareFn
Middleware Identifiers for each function

foreign_rpc_client
Helper module for jsonrpc clients. Automatically generated by easy-jsonrpc.

Traits

ForeignRpc
Public definition used to generate Foreign jsonrpc api.

OwnerRpc
Public definition used to generate Owner jsonrpc api.

Functions

run_doctest_foreign
helper to set up a real environment to run integrated doctests

run_doctest_owner
helper to set up a real environment to run integrated doctests

Type Definitions

ForeignCheckMiddleware
ForeignAPI Middleware Check callback